+++ src/main/java/capstone/safeat/oauth/external/GoogleOauthConfig.java
@@ -1,9 +1,9 @@
-package capstone.safeat.oauth;
+package capstone.safeat.oauth.external;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @ConfigurationProperties(prefix = "oauth.google")
-public record GoogleOAuthConfig(
+public record GoogleOauthConfig(
     String redirectUri,
     String clientId,
     String clientSecret
